Python/ReviewSession17/Poll/user/views.py
from rest_framework import viewsets
from user.serializers import UserSerializer
from django.contrib.auth.models import User


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
