Python/ReviewSession17/Poll/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
django==3.2.7
django-behaviors==0.5.1
django-environ==0.6.0
django-filter==2.4.0
djangorestframework==3.12.4
drf-jwt==1.19.1
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
markupsafe==2.0.1
packaging==21.0
pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
pyjwt[crypto]==2.1.0
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sentry-sdk==1.3.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
uwsgi==2.0.20
django-cors-headers==3.10.0
black